Add transcode symlinks only if binaries don't exist

+++ install/docker/run.sh
@@ -6,11 +6,11 @@ echo "Process will run as:"
 echo "User: $(id -u)"
 echo "Group: $(id -g)"
 mkdir -p $AIRSONIC_DIR/airsonic/transcode
-ln -fs /usr/bin/ffmpeg $AIRSONIC_DIR/airsonic/transcode/ffmpeg
-ln -fs /usr/bin/lame $AIRSONIC_DIR/airsonic/transcode/lame
+[[ ! -f $AIRSONIC_DIR/airsonic/transcode/ffmpeg ]] && ln -fs /usr/bin/ffmpeg $AIRSONIC_DIR/airsonic/transcode/ffmpeg
+[[ ! -f $AIRSONIC_DIR/airsonic/transcode/lame ]] && ln -fs /usr/bin/lame $AIRSONIC_DIR/airsonic/transcode/lame
 java --version
 echo "JAVA_OPTS=$JAVA_OPTS"
-/usr/bin/ffmpeg -version
+$AIRSONIC_DIR/airsonic/transcode/ffmpeg -version
 curl --version
 echo "PATH=$PATH"
